// src/modules/common/helpers/enums/mobile-pages.enum.ts
export type MobileScreenEnum =
    | 'splashLogin'
    | 'login'
    | 'subscription'
    | 'forgotPassword'
    | 'feed'
    | 'search'
    | 'sectionCompanies'
    | 'company'
    | 'product'
    | 'suggestCompany'
    | 'profile'
    | 'favoriteProduct'
    | 'creditHistory'
    | 'companyOrders'
    | 'deliveries'
    | 'listDeliveries'
    | 'cart'
    | 'orderList'
    | 'order'
    | 'selectAddress'
    | 'newAddress'
    | 'map'
    | 'typeAddress';